requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework-stubs==3.16.2
djangorestframework_simplejwt==5.5.1
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.3
types-PyYAML==6.0.12.20250822
types-requests==2.32.4.20250809
typing_extensions==4.14.1
urllib3==2.5.0
